fix: LottoStatistics클래스의 메소드 LottoWinPercentage 클래스로 책임 이동
